+++ src/main/java/kz/ncanode/configuration/SystemConfiguration.java
@@ -1,4 +1,4 @@
-package kz.ncanode.configuration.properties;
+package kz.ncanode.configuration;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -9,7 +9,7 @@ import org.springframework.context.annotation.Configuration;
 @ConfigurationProperties(prefix = "ncanode.system")
 @Getter
 @Setter
-public class SystemConfigurationProperties {
+public class SystemConfiguration {
     private boolean detailedErrors;
     private String cacheDir;
 }
